requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
browser-cookie3==0.13.0
bs4==0.0.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
cryptography==36.0.1
cssselect==1.1.0
dateparser==1.0.0
demjson3==3.0.5
fake-useragent==0.1.11
idna==2.10
importlib-metadata==4.10.0
jeepney==0.7.1
kafka-python==2.0.2
keyring==23.4.0
lxml==4.6.2
lz4==3.1.10
parse==1.19.0
pbkdf2==1.3
pyaes==1.6.1
pycparser==2.21
pycryptodome==3.12.0
pyee==8.1.0
pyppeteer==0.0.25
pyquery==1.4.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
regex==2021.3.17
requests==2.25.1
requests-html==0.10.0
SecretStorage==3.3.1
six==1.15.0
soupsieve==2.2.1
tqdm==4.59.0
tzlocal==2.1
urllib3==1.26.4
w3lib==1.22.0
websockets==8.0.2
zipp==3.6.0
